fix(): rename DistinctNamespaces and move it to namespaces
